Only compute damage on fresh hit
With this the damage calculation only runs if the hit instance
is older than 200ms. HeroController.TakeDamage runs every time
a collision with an enemy/hazard is detected, but this can
happen multiple times during a single damage event. By returning
early in such cases we reduce the CPU load slightly